Fixes to cache benchmark
0. I think we are occasionally getting stuck due to the flush timeout problem bdmyers noted---I had to add flushes to get it to work at one point. However, it's working right now without flushes in the system code, so I'll leave it like that and fix the timeout problem.

1. I think maybe I forgot to store the thread to wake up in the releaser, so it was trying to wake up a null pointer and crashing. When I checked it in, the code was there, though---so maybe I introduced this during debugging and then re-fixed it.

2. When the ready queue was empty, a suspend would silently fail and no suspend would happen. Now, this might be okay in the long run, but for now, the polling thread should always be on the run queue, and thus if the run queue is empty, something bad happened. I added an assert there for now.

3. Why was the run queue empty? There were two issues here.

3a. The run queue was empty because we called a blocking operation from an active message. In this case, "blocking" means "suspends the current thread". Active messages are executed by the polling thread, so when we do a cache operation in an active message that wants to block, it tries to suspend the polling thread. This is bad news, because 1) the polling thread shouldn't suspend, because we need it to receive more messages, and 2) there may be nothing else on the run queue at that point.

So fine. I turned the remote worker into a thread that gets spawned by an active message. This is annoying because we need to give the thread its arguments, but its arguments are in a buffer held by the polling thread which will go away when the current poll finishes. So I'm making a copy of the arguments, which means mallocing and freeing. In the future, we should probably follow Mark's suggestion and spawn the thread with a global pointer to its argument, so we can just copy the pointer to the thread and reference/cache the args instead of copying.

Two things that were harder than I expected with this:
 - it would be nice to join on a remotely-spawned thread. This can be done today, but it requres a lot of code and maybe a local proxy thread to make it clean. If we allowed global pointers in the join list, we could make this easy.
 - I thought it would be a good idea to make all the arg struct pointers typesafe. (I just added a new spawn construct to help with that.) The problem with this is that there's a lot of request/reply args, but the request arg struct type needs to have a global pointer to a reply arg struct type, which probably needs a pointer to the reply active message type, which also needs a pointer to its arg struct type. so there's this annoying circular type dependence that requires more code than I'd like to break.

Anyway.

3b. The run queue was empty because user_main called SoftXMT_signal_done() while work was still happening. This killed the polling threads, and so we were (mostly) unable to receive any messages.

My intention was that user_main would wait until it knows all work in the system is done before it calls SoftXMT_signal_done(). All this does is a broadcast quit for the polling threads, allowing everybody to leave the user-level scheduler and proceed to the finish statement. The polling threads will continue to spin/yield until they see the done message, so it's necessary to tell them to exit eventually, but you don't want to do it while work is still happening (and I chose to leave this to the user for now rather than automating it, since it's hard to know when work is done, as bdmyers can tell us).

So in this case, we needed some way to wait until the work is done so that the aggregators would still be around to receive messages. I added a return message from the workers and a blocking wait in user_main.