Add feed-forward delegate ops that use GlobalTaskJoiner rather than suspending.
These are not totally general, but could be useful in cases where you'd
like to do a delegate operation that you don't need a return value from.
Note: doesn't actually do fewer messages because it has to do a
remoteSignal, but frees up task and allows for more outstanding
messages.